# lab_02/openapi.yaml
openapi: 3.0.3
info:
  title: Football Tournament Management
  version: 1.0.0
  description: API for app

# servers:
#   - url: https://api.example.com/v1
#     description: Main API server
...
paths:
  /users:
    post:
      tags:
        - user
      summary: Create a new user.
      description: Create a new user.
      requestBody:
        description: Create a new user.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/user"
        required: true
      responses:
        "200":
          description: OK.
        "400":
          description: Invalid input!
        "500":
          description: Connect to database failed!
  /users/{userid}:
    get:
      tags:
        - user
      summary: get user by id.
      description: get user by id.
      parameters:
        - name: id.
          in: path
          description: ID of user.
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/user"
        "400":
          description: id must be an integer.
        "404":
          description: user not found.

    put:
      tags:
        - user
      summary: update user.
      description: update user .
      parameters:
        - name: user.
          in: path
          description: instance of object User.
          required: true
          schema:
            $ref: "#/components/schemas/user"
      responses:
        "200":
          description: successfully
        "400":
          description: failed connect to DB.
    delete:
      tags:
        - user
      summary: delete user by id.
      description: delete user by id.
      parameters:
        - name: id.
          in: path
          description: ID of user.
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: successfully.
        "404":
          description: doesn't exists.

    patch:
      tags:
        - user
      parameters:
        - name: login.
          in: path
          required: true
          schema:
            type: string
        - name: password.
          in: path
          required: true
          schema:
            type: string
      summary: change information of user
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                login:
                  type: string
                password:
                  type: string
                name:
                  type: string
                role:
                  type: string

      responses:
        "200":
          description: success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/user"
        "400":
          description: invalid login
  /users/{userlogin}:
    get:
      tags:
        - user
      summary: get user by login.
      description: get user by login.
      parameters:
        - name: login.
          in: path
          description: login of user.
          required: true
          schema:
            type: string
      responses:
        "200":
          description: successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/user"
        "404":
          description: user not found.
  /users?userRole=string:
    get:
      tags:
        - user
      summary: get user by role.
      description: get user by role.
      parameters:
        - name: role.
          in: query
          description: role of user.
          required: true
          schema:
            type: string
      responses:
        "200":
          description: successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/user"
...
components:
  schemas:
    user:
      type: object
      properties:
        id:
          type: integer
          example: 1
        login:
          type: string
          example: admin
        password:
          type: string
          example: aaa
        role:
          type: string
          example: Referee
        name:
          type: string
          example: Andrey
...